JANITORIAL: Silence a lot of warnings.
These are a result of ARRAYSIZE returning an size_t and most client code
trying to compare signed variables against the ARRAYSIZE result.

For some cases one would be probably better off changing member variables to
unsigned, but that is a more dangerous change than casting.

+++ audio/midiparser_xmidi.cpp
@@ -110,10 +110,10 @@ void MidiParser_XMIDI::parseNextEvent(EventInfo &info) {
 		// Simplified XMIDI looping.
 		case 0x74: {	// XMIDI_CONTROLLER_FOR_LOOP
 				byte *pos = _position._play_pos;
-				if (_loopCount < ARRAYSIZE(_loop) - 1)
+				if (_loopCount < (int)ARRAYSIZE(_loop) - 1)
 					_loopCount++;
 				else
-					warning("XMIDI: Exceeding maximum loop count %d", ARRAYSIZE(_loop));
+					warning("XMIDI: Exceeding maximum loop count %lu", ARRAYSIZE(_loop));
 
 				_loop[_loopCount].pos = pos;
 				_loop[_loopCount].repeat = info.basic.param2;
