+++ engines/gob/save/saveload_v2.cpp
@@ -273,7 +273,7 @@ SaveLoad_v2::~SaveLoad_v2() {
 const SaveLoad_v2::SaveFile *SaveLoad_v2::getSaveFile(const char *fileName) const {
 	fileName = stripPath(fileName);
 
-	for (int i = 0; i < ARRAYSIZE(_saveFiles); i++)
+	for (size_t i = 0; i < ARRAYSIZE(_saveFiles); i++)
 		if (!scumm_stricmp(fileName, _saveFiles[i].sourceName))
 			return &_saveFiles[i];
 
@@ -283,7 +283,7 @@ const SaveLoad_v2::SaveFile *SaveLoad_v2::getSaveFile(const char *fileName) cons
 SaveLoad_v2::SaveFile *SaveLoad_v2::getSaveFile(const char *fileName) {
 	fileName = stripPath(fileName);
 
-	for (int i = 0; i < ARRAYSIZE(_saveFiles); i++)
+	for (size_t i = 0; i < ARRAYSIZE(_saveFiles); i++)
 		if (!scumm_stricmp(fileName, _saveFiles[i].sourceName))
 			return &_saveFiles[i];
