+++ engines/saga/scene.cpp
@@ -452,7 +452,7 @@ void Scene::changeScene(int16 sceneNumber, int actorsEntrance, SceneTransitionTy
 	// This is used for latter ITE demos where all places on world map except
 	// Tent Faire are substituted with LBM picture and short description
 	if (_vm->_hasITESceneSubstitutes) {
-		for (int i = 0; i < ARRAYSIZE(sceneSubstitutes); i++) {
+		for (size_t i = 0; i < ARRAYSIZE(sceneSubstitutes); i++) {
 			if (sceneSubstitutes[i].sceneId == sceneNumber) {
 				Surface bbmBuffer;
 				byte *pal, colors[768];
